Change naming to fit documentation
