+++ climbers-application/src/main/java/com/dostojic/climbers/logic/so/climber/DeleteClimber.java
@@ -14,11 +14,11 @@ import com.dostojic.climbers.logic.so.template.GeneralUpdateSO;
  *
  * @author Dejan.Ostojic
  */
-public class DeleteClimberSO extends GeneralUpdateSO<Integer, Boolean> {
+public class DeleteClimber extends GeneralUpdateSO<Integer, Boolean> {
 
     private final ClimberRepository climberRepository;
 
-    public DeleteClimberSO(TransactionManager transactionManager, ClimberRepository climberRepository) {
+    public DeleteClimber(TransactionManager transactionManager, ClimberRepository climberRepository) {
         super(transactionManager);
         this.climberRepository = climberRepository;
     }
