+++ climbers-application/src/main/java/com/dostojic/climbers/logic/so/climber/FindClimber.java
@@ -16,11 +16,11 @@ import java.util.Optional;
  *
  * @author Dejan.Ostojic
  */
-public class GetClimberDetailsSO extends GeneralReportingSO<Integer, Climber>{
+public class FindClimber extends GeneralReportingSO<Integer, Climber>{
 
     private final ClimberRepository climberRepository;
     
-    public GetClimberDetailsSO(TransactionManager transactionManager, ClimberRepository climberRepository) {
+    public FindClimber(TransactionManager transactionManager, ClimberRepository climberRepository) {
         super(transactionManager);
         this.climberRepository = climberRepository;
     }
