+++ climbers-application/src/main/java/com/dostojic/climbers/logic/so/template/GeneralUpdateSO.java
@@ -10,8 +10,6 @@ import com.dostojic.climbers.logic.TransactionManager;
 /**
  *
  * @author Dejan.Ostojic
- * @param <T>
- * @param <DomainObject>
  */
 public abstract class GeneralUpdateSO<DomainObjectIn, DomainObjectOut>  implements GeneralSO<DomainObjectIn, DomainObjectOut> {
