+++ climbers-repository-dbbr-adapter/src/main/resources/hikariConfig.properties
@@ -1,6 +1,6 @@
-jdbcUrl=jdbc:mysql://localhost:3306/ps_climbers
+jdbcUrl=jdbc:mysql://localhost:3306/ps_climbers?useLegacyDatetimeCode=false&serverTimezone=UTC
 username=root
-password=root
+password=
 dataSource.cachePrepStmts=true
 dataSource.prepStmtCacheSize=250
 dataSource.prepStmtCacheSqlLimit=2048
