+++ climbers-repository-dbbr-adapter/src/test/resources/hikariConfig.properties
@@ -1,6 +1,6 @@
 jdbcUrl=jdbc:h2:mem:test;DB_CLOSE_DELAY=-1;INIT=runscript from 'classpath:/schema.sql'
 username=root
-password=root
+password=
 dataSource.cachePrepStmts=true
 dataSource.prepStmtCacheSize=250
 dataSource.prepStmtCacheSqlLimit=2048
